Extracted LicenseLabel component
